Move Manifest in its own module file (and test)
